reduce max file size
