+++ enclone_print/src/print_clonotypes.rs
@@ -701,300 +701,36 @@ pub fn print_clonotypes(
                     }
                     res.5 = gex_low;
 
-                    // Very bad computation because of embedded binary search.
+                    // Compute per-cell entries.
 
                     let mut subrows = Vec::<Vec<String>>::new();
-                    if ctl.clono_print_opt.bu {
-                        for bcl in bli.iter() {
-                            let mut row = Vec::<String>::new();
-                            let bc = &bcl.0;
-                            let li = bcl.1;
-                            let di = ex.clones[bcl.2][0].dataset_index;
-                            row.push(format!("$  {}", bc.clone()));
-                            let ex = &exact_clonotypes[exacts[u]];
-                            for k in 0..lvars.len() {
-                                let nr = row.len();
-                                let mut filled = false;
-                                for l in 0..ctl.origin_info.n() {
-                                    if lvars[k] == format!("n_{}", ctl.origin_info.dataset_id[l]) {
-                                        let mut n = 0;
-                                        if di == l {
-                                            n = 1;
-                                        }
-                                        row.push(format!("{}", n));
-                                        filled = true;
-                                    }
-                                }
-                                if filled {
-                                } else if lvars[k] == "n_b".to_string() {
-                                    let mut n = 0;
-                                    let li = ex.clones[bcl.2][0].dataset_index;
-                                    if gex_info.cell_type[li].contains_key(&bc.clone()) {
-                                        if gex_info.cell_type[li][&bc.clone()].starts_with('B') {
-                                            n = 1;
-                                        }
-                                    }
-                                    row.push(format!("{}", n));
-                                } else if lvars[k] == "filter".to_string() {
-                                    let mut f = String::new();
-                                    if fate[li].contains_key(&bc.clone()) {
-                                        f = fate[li][&bc.clone()].clone();
-                                        f = f.between(" ", " ").to_string();
-                                    }
-                                    row.push(f);
-                                } else if lvars[k] == "n_other".to_string() {
-                                    let mut n = 0;
-                                    let di = ex.clones[bcl.2][0].dataset_index;
-                                    let f = format!("n_{}", ctl.origin_info.dataset_id[di]);
-                                    let mut found = false;
-                                    for i in 0..nd_fields.len() {
-                                        if f == nd_fields[i] {
-                                            found = true;
-                                        }
-                                    }
-                                    if !found {
-                                        n = 1;
-                                    }
-                                    row.push(format!("{}", n));
-                                } else if lvars[k] == "sec".to_string() {
-                                    let mut n = 0;
-                                    if ctl.origin_info.secmem[li].contains_key(&bc.clone()) {
-                                        n = ctl.origin_info.secmem[li][&bc.clone()].0;
-                                    }
-                                    row.push(format!("{}", n));
-                                } else if lvars[k] == "mem".to_string() {
-                                    let mut n = 0;
-                                    if ctl.origin_info.secmem[li].contains_key(&bc.clone()) {
-                                        n = ctl.origin_info.secmem[li][&bc.clone()].1;
-                                    }
-                                    row.push(format!("{}", n));
-                                } else if bin_member(&alt_bcs, &lvars[k]) {
-                                    let mut val = String::new();
-                                    let alt = &ctl.origin_info.alt_bc_fields[li];
-                                    for j in 0..alt.len() {
-                                        if alt[j].0 == lvars[k] {
-                                            if alt[j].1.contains_key(&bc.clone()) {
-                                                val = alt[j].1[&bc.clone()].clone();
-                                            }
-                                        }
-                                    }
-                                    row.push(val);
-                                } else if lvars[k] == "datasets".to_string() {
-                                    row.push(format!("{}", ctl.origin_info.dataset_id[li].clone()));
-                                } else if lvars[k] == "clust".to_string() && have_gex {
-                                    let mut cid = 0;
-                                    if gex_info.cluster[li].contains_key(&bc.clone()) {
-                                        cid = gex_info.cluster[li][&bc.clone()];
-                                    }
-                                    row.push(format!("{}", cid));
-                                } else if lvars[k].starts_with("pe") && have_gex {
-                                    row.push(format!("{}", pe[k][cell_count + bcl.2]));
-                                } else if lvars[k].starts_with("npe") && have_gex {
-                                    row.push(format!("{}", npe[k][cell_count + bcl.2]));
-                                } else if lvars[k].starts_with("ppe") && have_gex {
-                                    row.push(format!("{}", ppe[k][cell_count + bcl.2]));
-                                } else if lvars[k] == "cred".to_string() && have_gex {
-                                    row.push(format!("{}", cred[k][cell_count + bcl.2]));
-                                } else if lvars[k] == "type".to_string() && have_gex {
-                                    let mut cell_type = "".to_string();
-                                    if gex_info.cell_type[li].contains_key(&bc.clone()) {
-                                        cell_type = gex_info.cell_type[li][&bc.clone()].clone();
-                                    }
-                                    row.push(cell_type);
-                                } else if lvars[k] == "n_gex".to_string() && have_gex {
-                                    let mut n_gex = 0;
-                                    if bin_member(&gex_info.gex_cell_barcodes[li], &bc) {
-                                        n_gex = 1;
-                                    }
-                                    row.push(format!("{}", n_gex));
-                                } else if lvars[k] == "mark".to_string() {
-                                    let mut mark = String::new();
-                                    if ex.clones[bcl.2][0].marked {
-                                        mark = "x".to_string();
-                                    }
-                                    row.push(mark);
-                                } else if lvars[k] == "entropy".to_string() && have_gex {
-                                    // NOTE DUPLICATION WITH CODE BELOW.
-                                    let mut gex_count = 0;
-                                    let p = bin_position(&gex_info.gex_barcodes[li], &bc);
-                                    if p >= 0 {
-                                        let mut raw_count = 0;
-                                        if gex_info.gex_matrices[li].initialized() {
-                                            let row = gex_info.gex_matrices[li].row(p as usize);
-                                            for j in 0..row.len() {
-                                                let f = row[j].0;
-                                                let n = row[j].1;
-                                                if gex_info.is_gex[li][f] {
-                                                    raw_count += n;
-                                                }
-                                            }
-                                        } else {
-                                            let l = bcl.2;
-                                            for j in 0..d_all[l].len() {
-                                                if gex_info.is_gex[li][ind_all[l][j] as usize] {
-                                                    raw_count += d_all[l][j] as usize;
-                                                }
-                                            }
-                                        }
-                                        gex_count = raw_count;
-                                    }
-                                    let mut entropy = 0.0;
-                                    if p >= 0 {
-                                        if gex_info.gex_matrices[li].initialized() {
-                                            let row = gex_info.gex_matrices[li].row(p as usize);
-                                            for j in 0..row.len() {
-                                                let f = row[j].0;
-                                                let n = row[j].1;
-                                                if gex_info.is_gex[li][f] {
-                                                    let q = n as f64 / gex_count as f64;
-                                                    entropy -= q * q.log2();
-                                                }
-                                            }
-                                        } else {
-                                            let l = bcl.2;
-                                            for j in 0..d_all[l].len() {
-                                                if gex_info.is_gex[li][ind_all[l][j] as usize] {
-                                                    let n = d_all[l][j] as usize;
-                                                    let q = n as f64 / gex_count as f64;
-                                                    entropy -= q * q.log2();
-                                                }
-                                            }
-                                        }
-                                    }
-                                    row.push(format!("{:.2}", entropy));
-                                } else if have_gex {
-                                    // this calc isn't needed except in _% case below
-                                    // TODO: ELIMINATE UNNEEDED CALC
-                                    let mut gex_count = 0.0;
-                                    let p = bin_position(&gex_info.gex_barcodes[li], &bc);
-                                    if p >= 0 {
-                                        let mut raw_count = 0 as f64;
-                                        if gex_info.gex_matrices[li].initialized() {
-                                            let row = gex_info.gex_matrices[li].row(p as usize);
-                                            for j in 0..row.len() {
-                                                let f = row[j].0;
-                                                let n = row[j].1;
-                                                if gex_info.is_gex[li][f] {
-                                                    raw_count += n as f64;
-                                                }
-                                            }
-                                        } else {
-                                            let l = bcl.2;
-                                            for j in 0..d_all[l].len() {
-                                                if gex_info.is_gex[li][ind_all[l][j] as usize] {
-                                                    raw_count += d_all[l][j] as f64;
-                                                }
-                                            }
-                                        }
-                                        if !ctl.gen_opt.full_counts {
-                                            gex_count = raw_count * gex_info.gex_mults[li];
-                                        } else {
-                                            gex_count = raw_count;
-                                        }
-                                    }
-                                    if lvars[k] == "gex".to_string() {
-                                        row.push(format!("{}", gex_count.round()));
-                                    } else {
-                                        let mut y = lvars[k].clone();
-                                        if y.contains(':') {
-                                            y = y.after(":").to_string();
-                                        }
-                                        let y0 = y.clone();
-                                        let suffixes = ["_min", "_max", "_μ", "_Σ", "_cell", "_%"];
-                                        for s in suffixes.iter() {
-                                            if y.ends_with(s) {
-                                                y = y.rev_before(&s).to_string();
-                                                break;
-                                            }
-                                        }
-                                        let p = bin_position(&gex_info.gex_barcodes[li], &bc);
-                                        let mut computed = false;
-                                        let mut count = 0.0;
-                                        let l = bcl.2;
-                                        if p >= 0 {
-                                            let mut ux = Vec::<usize>::new();
-                                            if ctl.clono_print_opt.regex_match[li].contains_key(&y)
-                                            {
-                                                ux =
-                                                    ctl.clono_print_opt.regex_match[li][&y].clone();
-                                            }
-                                            if ux.len() > 0 {
-                                                computed = true;
-                                                for fid in ux.iter() {
-                                                    let counti = get_gex_matrix_entry(
-                                                        &ctl, &gex_info, *fid, &d_all, &ind_all,
-                                                        li, l, p as usize, &y,
-                                                    );
-                                                    count += counti;
-                                                }
-                                            } else if gex_info.feature_id[li].contains_key(&y) {
-                                                computed = true;
-                                                let fid = gex_info.feature_id[li][&y];
-                                                count = get_gex_matrix_entry(
-                                                    &ctl, &gex_info, fid, &d_all, &ind_all, li, l,
-                                                    p as usize, &y,
-                                                );
-                                            }
-                                        }
-                                        if computed {
-                                            // note unneeded calculation above in certain cases
-                                            // TODO: ELIMINATE!
-                                            if y0.ends_with("_min") {
-                                            } else if y0.ends_with("_max") {
-                                            } else if y0.ends_with("_μ") {
-                                            } else if y0.ends_with("_Σ") {
-                                            } else if y0.ends_with("_%") {
-                                                row.push(format!(
-                                                    "{:.2}",
-                                                    (100.0 * count) / gex_count
-                                                ));
-                                            } else {
-                                                row.push(format!("{}", count.round()));
-                                            }
-                                        }
-                                    }
-                                }
-                                if row.len() == nr {
-                                    row.push("".to_string());
-                                }
-                            }
-                            let mut ncall = 0;
-                            for k in 0..cols {
-                                ncall += rsi.cvars[k].len();
-                            }
-                            let mut cx = vec!["".to_string(); ncall];
-                            let mut cp = 0;
-                            for col in 0..cols {
-                                let m = mat[col][u];
-                                if m.is_some() {
-                                    let m = m.unwrap();
-                                    for p in 0..rsi.cvars[col].len() {
-                                        if rsi.cvars[col][p] == "u".to_string() {
-                                            let numi = ex.clones[bcl.2][m].umi_count;
-                                            cx[cp + p] = format!("{}", numi);
-                                        } else if rsi.cvars[col][p] == "r".to_string() {
-                                            let r = ex.clones[bcl.2][m].read_count;
-                                            cx[cp + p] = format!("{}", r);
-                                        } else if rsi.cvars[col][p] == "nval".to_string() {
-                                            let mut n = 0;
-                                            if ex.clones[bcl.2][m].validated_umis.is_some() {
-                                                n = ex.clones[bcl.2][m]
-                                                    .validated_umis
-                                                    .as_ref()
-                                                    .unwrap()
-                                                    .len();
-                                            }
-                                            cx[cp + p] = format!("{}", n);
-                                        }
-                                    }
-                                }
-                                cp += rsi.cvars[col].len();
-                            }
-                            row.append(&mut cx);
-                            subrows.push(row);
-                        }
-                    }
-                    sr.push((row, subrows, varmat[u].clone(), u));
+                    compute_bu(
+                        u,
+                        cell_count,
+                        &exacts,
+                        &lvars,
+                        &ctl,
+                        &bli,
+                        &ex,
+                        &exact_clonotypes,
+                        &mut row,
+                        &mut subrows,
+                        &varmat,
+                        have_gex,
+                        &gex_info,
+                        &rsi,
+                        &mut sr,
+                        &fate,
+                        &nd_fields,
+                        &alt_bcs,
+                        &cred,
+                        &pe,
+                        &ppe,
+                        &npe,
+                        &d_all,
+                        &ind_all,
+                        &mat,
+                    );
                     cell_count += ex.clones.len();
                 }
                 let mut rord = Vec::<usize>::new(); // note that this is now superfluous
